moved corrected log filtering to util folder
